+++ examples/resources/catalog_entity/resource.tf
@@ -309,6 +309,19 @@ resource "cortex_catalog_entity" "products-service" {
     ]
   }
 
+  launch_darkly = {
+    projects = [
+      {
+        id   = "products-service"
+        type = "KEY"
+        environments = [
+          { name : "staging" },
+          { name : "production" },
+        ]
+      }
+    ]
+  }
+
   microsoft_teams = [
     {
       name                  = "engineering"